Revert "Memory optimizing"
This reverts commit 10e5b1c530661c34ec2f5f96c85672a9e0b42c6d.

+++ BattleCity/BattleCity/GameObject.cpp
@@ -1,6 +1,6 @@
 #include "GameObject.h"
 
-GameObject::GameObject(uint8_t xStart, uint8_t yStart, int id, bool isBreakable, bool isVisible)
+GameObject::GameObject(int xStart, int yStart, int id, bool isBreakable, bool isVisible)
     : m_x{ xStart }, 
     m_y{ yStart }, 
     m_id{ id },
